# docs/swagger.yaml
basePath: /api
definitions:
...
  request.GroupAddReq:
    properties:
      groupName:
        maxLength: 128
        minLength: 1
        type: string
      groupType:
        maxLength: 20
        minLength: 1
        type: string
      parentId:
        description: 父级Id 大于等于0 必填
        minimum: 0
        type: integer
      remark:
        description: 分组的中文描述
        maxLength: 128
        minLength: 0
        type: string
    required:
    - groupName
    - groupType
    type: object
  request.GroupAddUserReq:
    properties:
      groupId:
        type: integer
      userIds:
        items:
          type: integer
        type: array
    required:
    - groupId
    - userIds
    type: object
  request.GroupDeleteReq:
    properties:
      groupIds:
        items:
          type: integer
        type: array
    required:
    - groupIds
    type: object
  request.GroupRemoveUserReq:
    properties:
      groupId:
        type: integer
      userIds:
        items:
          type: integer
        type: array
    required:
    - groupId
    - userIds
    type: object
  request.GroupUpdateReq:
    properties:
      groupName:
        maxLength: 128
        minLength: 1
        type: string
      id:
        type: integer
      remark:
        description: 分组的中文描述
        maxLength: 128
        minLength: 0
        type: string
    required:
    - groupName
    - id
    type: object
  request.MenuAddReq:
    properties:
      activeMenu:
        maxLength: 100
        minLength: 0
        type: string
      alwaysShow:
        enum:
          - 1
          - 2
        type: integer
      breadcrumb:
        enum:
          - 1
          - 2
        type: integer
      component:
        maxLength: 100
        minLength: 1
        type: string
      hidden:
        enum:
          - 1
          - 2
        type: integer
      icon:
        maxLength: 50
        minLength: 0
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
      noCache:
        enum:
          - 1
          - 2
        type: integer
      parentId:
        type: integer
      path:
        maxLength: 100
        minLength: 1
        type: string
      redirect:
        maxLength: 100
        minLength: 0
        type: string
      sort:
        maximum: 999
        minimum: 1
        type: integer
      status:
        enum:
          - 1
          - 2
        type: integer
      title:
        maxLength: 50
        minLength: 1
        type: string
    required:
      - component
      - name
      - path
      - title
    type: object
  request.MenuDeleteReq:
    properties:
      menuIds:
        items:
          type: integer
        type: array
    required:
      - menuIds
    type: object
  request.MenuUpdateReq:
    properties:
      activeMenu:
        maxLength: 100
        minLength: 0
        type: string
      alwaysShow:
        enum:
          - 1
          - 2
        type: integer
      breadcrumb:
        enum:
          - 1
          - 2
        type: integer
      component:
        maxLength: 100
        minLength: 0
        type: string
      hidden:
        enum:
          - 1
          - 2
        type: integer
      icon:
        maxLength: 50
        minLength: 0
        type: string
      id:
        type: integer
      name:
        maxLength: 50
        minLength: 1
        type: string
      noCache:
        enum:
          - 1
          - 2
        type: integer
      parentId:
        minimum: 0
        type: integer
      path:
        maxLength: 100
        minLength: 1
        type: string
      redirect:
        maxLength: 100
        minLength: 0
        type: string
      sort:
        maximum: 999
        minimum: 1
        type: integer
      status:
        enum:
          - 1
          - 2
        type: integer
      title:
        maxLength: 50
        minLength: 1
        type: string
    required:
      - id
      - name
      - path
      - title
    type: object
  request.OperationLogDeleteReq:
    properties:
      operationLogIds:
        items:
          type: integer
        type: array
    required:
      - operationLogIds
    type: object
...
  response.ResponseBody:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.1:8888
info:
  contact:
    email: contact@example.com
    name: Example
    url: https://example.com
  description: 基于Go+Vue实现的openLDAP后台管理项目
  termsOfService: https://example.com
  title: Go Ldap Admin
  version: "1.0"
paths:
...
  /group/add:
    post:
      consumes:
      - application/json
      description: 添加分组记录
      parameters:
      - description: 添加用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 添加分组记录
      tags:
      - 分组管理
  /group/adduser:
    post:
      consumes:
      - application/json
      description: 添加用户
      parameters:
      - description: 添加用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupAddUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 添加用户
      tags:
      - 分组管理
  /group/delete:
    post:
      consumes:
      - application/json
      description: 删除分组记录
      parameters:
      - description: 删除用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 删除分组记录
      tags:
      - 分组管理
  /group/list:
    get:
      consumes:
      - application/json
      description: 获取分组记录列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 获取分组记录列表
      tags:
      - 分组管理
  /group/removeuser:
    post:
      consumes:
      - application/json
      description: 移除用户
      parameters:
      - description: 移除用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupRemoveUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 移除用户
      tags:
      - 分组管理
  /group/syncDingTalkDepts:
    post:
      consumes:
      - application/json
      description: 同步钉钉部门信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 同步钉钉部门信息
      tags:
      - 分组管理
  /group/syncFeiShuDepts:
    post:
      consumes:
      - application/json
      description: 同步飞书部门信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 同步飞书部门信息
      tags:
      - 分组管理
  /group/syncOpenLdapDepts:
    post:
      consumes:
      - application/json
      description: 同步原ldap部门信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 同步原ldap部门信息
      tags:
      - 分组管理
  /group/syncSqlGroups:
    post:
      consumes:
      - application/json
      description: 同步Sql中的分组信息到ldap
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 同步Sql中的分组信息到ldap
      tags:
      - 分组管理
  /group/syncWeComDepts:
    post:
      consumes:
      - application/json
      description: 同步企业微信部门信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 同步企业微信部门信息
      tags:
      - 分组管理
  /group/tree:
    get:
      consumes:
      - application/json
      description: 获取分组接口树
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 获取分组接口树
      tags:
      - 分组管理
  /group/update:
    post:
      consumes:
      - application/json
      description: 更新分组记录
      parameters:
      - description: 更新用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 更新分组记录
      tags:
      - 分组管理
  /group/useringroup:
    get:
      consumes:
      - application/json
      description: 获取分组内用户
      parameters:
      - description: 分组ID
        in: query
        name: groupId
        required: true
        type: integer
      - description: 昵称
        in: query
        name: nickname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 获取分组内用户
      tags:
      - 分组管理
  /group/usernoingroup:
    get:
      consumes:
      - application/json
      description: 不在分组的用户
      parameters:
      - description: 分组ID
        in: query
        name: groupId
        required: true
        type: integer
      - description: 昵称
        in: query
        name: nickname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 不在分组的用户
      tags:
      - 分组管理
  /log/operation/clean:
    delete:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 清空操作日志记录
      tags:
        - 操作日志管理
  /log/operation/delete:
    post:
      consumes:
        - application/json
      parameters:
        - description: 删除日志的ID
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.OperationLogDeleteReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 删除操作日志记录
      tags:
        - 操作日志管理
  /log/operation/list:
    get:
      consumes:
        - application/json
      parameters:
        - description: 用户名
          in: query
          name: username
          type: string
        - description: IP地址
          in: query
          name: ip
          type: string
        - description: 路径
          in: query
          name: path
          type: string
        - description: 方法
          in: query
          name: method
          type: string
        - description: 状态码
          in: query
          name: status
          type: integer
        - description: 页码
          in: query
          name: pageNum
          type: integer
        - description: 每页数量
          in: query
          name: pageSize
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 获取操作日志记录列表
      tags:
        - 操作日志管理
  /menu/access/tree:
    get:
      consumes:
        - application/json
      parameters:
        - description: 分组ID
          in: query
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 获取用户菜单树
      tags:
        - 菜单管理
  /menu/add:
    post:
      consumes:
        - application/json
      parameters:
        - description: 新建菜单
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.MenuAddReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 新建菜单
      tags:
        - 菜单管理
  /menu/delete:
    post:
      consumes:
        - application/json
      parameters:
        - description: 删除菜单
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.MenuDeleteReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 删除菜单
      tags:
        - 菜单管理
  /menu/tree:
    get:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 获取菜单树
      tags:
        - 菜单管理
  /menu/update:
    post:
      consumes:
        - application/json
      parameters:
        - description: 更新菜单
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.MenuUpdateReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 更新菜单
      tags:
        - 菜单管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"